Fix Caching Issues
The python userbase in the setup python action was not being cached due
to the path used being incorrect. I seem to recall finding the path in
the setup-python or caching action docs, however I can no longer find
it. It's been removed as it's not very relevant either way.

The pip cache has been added to the set of cached paths.
The tests' python patch version has been specified to avoid
other cache misses.

Signed-off-by: Hassan Abouelela <hassan@hassanamr.com>